perf(day8): combine parts 1/2.
I don't usually like relying on the mutability of the solve, but in this
case, the argument is that it's faster to solve both parts in the slow
way than it would be to solve each part individually in a "fast" way. I
already permit for mutability where part two needs to build off of the
result from part one, and, technically, calculating how far we can see
in each direction to determine if we can see an edge _is_ the result
from part one.