perf(day16): huge preformance boost just looking for disjoint paths.
Because my solution was just letting the elephant work every time we ran
out of time, it was effectively restricting to the valves I hadn't
opened yet. So we can just loop through the set of all paths looking for
disjoint ones and summing their totals, finding the best combination. It
seems obvious now, but clearly took me two days to think up.